refactor: extract YampGridLayout into a dedicated module to improve codebase modularity
